third_party/blink/web_tests/fast/dom/HTMLImageElement/image-srcset-in-iframe-expected.txt

PASS internals.isLoadingFromMemoryCache("resources/image-set-4x.png") is false
PASS internals.isLoadingFromMemoryCache("resources/blue_rect.jpg") is true
PASS successfullyParsed is true

TEST COMPLETE
